Adds keySize parameter to AES functions
Adds a `keySize` parameter to each of the AES encryption/decryption functions. The parameter defaults to kCCKeySizeAES128, which is what was being used previously.
Also creates appropariately sized hashedKeyData and ivData based on the specified key size. (I believe the previous code was incorrecly creating a 256-bit key data, when only specifying 128-bit key size in the CCCrypt() options.)